+++ Engine/Engine.Main/Include/Public/Common/PonyEngine/Engine/EngineModule.h
@@ -1,0 +1,21 @@
+/***************************************************
+ * MIT License                                     *
+ *                                                 *
+ * Copyright (c) 2023-present Vladimir Popov       *
+ *                                                 *
+ * Email: zor1994@gmail.com                        *
+ * Repo: https://github.com/ZorPastaman/PonyEngine *
+ ***************************************************/
+
+#pragma once
+
+#include "PonyEngine/Utility/Compiler.h"
+
+import PonyEngine.Application;
+
+namespace PonyEngine::Engine
+{
+	/// @brief Gets the engine module.
+	/// @return Engine module.
+	PONY_DLL_EXPORT Application::IModule* GetEngineModule();
+}