+++ Engine/Log.Main/Source/Main.Impl.cpp
@@ -9,14 +9,14 @@
 
 #include "PonyEngine/Log/LoggerModule.h"
 
-import PonyEngine.Core;
+import PonyEngine.Application;
 import PonyEngine.Log.Main;
 
 namespace PonyEngine::Log
 {
 	LoggerModule LoggerModule; ///< Logger module.
 
-	Core::IModule* GetLoggerModule()
+	Application::IModule* GetLoggerModule()
 	{
 		return &LoggerModule;
 	}