+++ Engine/Log.Main/Source/Main.cpp
@@ -7,7 +7,7 @@
  * Repo: https://github.com/ZorPastaman/PonyEngine *
  ***************************************************/
 
-#include "PonyEngine/Core/Module.h"
+#include "PonyEngine/Application/Module.h"
 #include "PonyEngine/Log/LoggerModule.h"
 
-PONY_MODULE_LOG(PonyEngine::Log::GetLoggerModule, PonyEngineLogger);
+PONY_MODULE(PonyEngine::Log::GetLoggerModule, PonyEngineLogger, PONY_ENGINE_LOG_MAIN_ORDER);