+++ Game/EngineSystemSample/GameSample/Source/Main.Impl.cpp
@@ -1,0 +1,14 @@
+#include "Game/GameModule.h"
+
+import Game;
+
+namespace Game
+{
+	// The module lifetime must be the same as the library lifetime. So, it's much easier to define it as a global variable.
+	GameSystemModule GameModule;
+
+	PonyEngine::Application::IModule* GetGameModule()
+	{
+		return &GameModule;
+	}
+}