+++ CHANGELOG.md
@@ -12,10 +12,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Core module.
 - Math module.
 - Application main module.
+- Path module.
+- Path main module.
 - Log main module.
+- Log extension module.
 - Log console module.
 - Log platform console module.
 - Log file module.
+- Engine extension module.
 - Engine main module.
 
 [Unreleased]: https://github.com/ZorPastaman/PonyEngine